Deleted unused directories, renamed directories, and updated data collection code
